@@ -18,18 +18,24 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_dotenv: ^3.1.0
-  photo_view: ^0.10.3
+  flutter_riverpod: ^0.14.0+3
+  flutter_dotenv: ^5.0.2
+  photo_view: ^0.13.0
+  http: ^0.13.4
+  intl: ^0.17.0
+  path: ^1.8.0
+  path_provider: ^2.0.5
+  gallery_saver: ^2.3.1
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
